Add user specs pending_friends and friend_requests

+++ app/controllers/likes_controller.rb
@@ -5,10 +5,8 @@ class LikesController < ApplicationController
   end
 
   def destroy
-    like = Like.find_by(id: params[:id], user: current_user, post_id: param[:post_id])
-    if like
-      like.destroy
-      redirect_to posts_path, notice: 'You disliked a post.'
-    end
+    like = Like.find_by(id: params[:id], user: current_user, post_id: params[:post_id])
+    like.destroy if like
+    redirect_to posts_path, notice: 'You disliked a post.'
   end
 end
